+++ keras/layers/__init__.py
@@ -2,6 +2,7 @@ from __future__ import absolute_import
 from ..engine import Layer, Input, InputLayer, Merge, merge, InputSpec
 from .core import *
 from .convolutional import *
+from .pooling import *
 from .recurrent import *
 from .normalization import *
 from .embeddings import *
